Moved Build History pagination and search classes to jenkins as requested by @jglick
I understand the motivation for this but I think it creates a bit of a maint/comprehension issue in that splits related classes across 2 different package hierarchies. I know we'll say that's ok but I don't like it. Just say'n.
