+++ core/src/main/java/jenkins/widgets/buildsearch/BuildSearchParams.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package hudson.widgets.buildsearch;
+package jenkins.widgets.buildsearch;
 
 import javax.annotation.Nonnull;
 import java.util.ArrayList;
