+++ core/src/main/java/jenkins/widgets/buildsearch/README.md
