+++ core/src/main/java/jenkins/widgets/buildsearch/processors/DescriptionProcessorFactory.java
@@ -21,29 +21,29 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package hudson.widgets.buildsearch.processors;
+package jenkins.widgets.buildsearch.processors;
 
 import hudson.model.Queue;
 import hudson.model.Run;
-import hudson.widgets.buildsearch.BuildSearchParamProcessor;
-import hudson.widgets.buildsearch.BuildSearchParamProcessorFactory;
-import hudson.widgets.buildsearch.BuildSearchParams;
+import jenkins.widgets.buildsearch.BuildSearchParamProcessor;
+import jenkins.widgets.buildsearch.BuildSearchParamProcessorFactory;
+import jenkins.widgets.buildsearch.BuildSearchParams;
 
 import java.util.List;
 
 /**
- * Search build history by {@link hudson.model.Queue.Item} or {@link hudson.model.Run} name.
+ * Search build history by {@link hudson.model.Run} description.
  *
  * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
  */
-public class NameProcessorFactory extends BuildSearchParamProcessorFactory {
+public class DescriptionProcessorFactory extends BuildSearchParamProcessorFactory {
 
     /**
-     * "name" search term.
+     * "desc" (description) search term.
      */
-    public static final String NAME_ST = "name";
+    public static final String DESC_ST = "desc";
 
-    private static final String[] SEARCH_TERMS = new String[] {NAME_ST}; // Build name
+    private static final String[] SEARCH_TERMS = new String[] {DESC_ST}; // Build description
 
     /**
      * {@inheritDoc}
@@ -58,20 +58,20 @@ public class NameProcessorFactory extends BuildSearchParamProcessorFactory {
      */
     @Override
     public BuildSearchParamProcessor createProcessor(BuildSearchParams searchParams) {
-        final List<BuildSearchParams.BuildSearchParam> nameParams = searchParams.getParams(NAME_ST);
+        final List<BuildSearchParams.BuildSearchParam> descParams = searchParams.getParams(DESC_ST);
 
-        if (nameParams.isEmpty()) {
-            // "name" search term not specified in search
+        if (descParams.isEmpty()) {
+            // "desc" search term not specified in search
             return null;
         }
 
         return new BuildSearchParamProcessor<String>() {
 
             @Override
-            public boolean fitsSearchParams(String name) {
-                // It fits if it contains any of the specified "name" search terms.
-                for (BuildSearchParams.BuildSearchParam nameParam : nameParams) {
-                    if (name.contains(nameParam.get())) {
+            public boolean fitsSearchParams(String description) {
+                // It fits if it contains any of the specified "desc" search terms.
+                for (BuildSearchParams.BuildSearchParam nameParam : descParams) {
+                    if (description.contains(nameParam.get())) {
                         return true;
                     }
                 }
@@ -80,11 +80,12 @@ public class NameProcessorFactory extends BuildSearchParamProcessorFactory {
 
             @Override
             public boolean fitsSearchParams(Queue.Item item) {
-                return fitsSearchParams(item.getDisplayName());
+                // Queue items don't have a description.
+                return false;
             }
             @Override
             public boolean fitsSearchParams(Run run) {
-                return fitsSearchParams(run.getDisplayName());
+                return fitsSearchParams(run.getDescription());
             }
         };
     }
