+++ core/src/main/java/jenkins/widgets/buildsearch/processors/NameProcessorFactory.java
@@ -21,29 +21,29 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package hudson.widgets.buildsearch.processors;
+package jenkins.widgets.buildsearch.processors;
 
 import hudson.model.Queue;
 import hudson.model.Run;
-import hudson.widgets.buildsearch.BuildSearchParamProcessor;
-import hudson.widgets.buildsearch.BuildSearchParamProcessorFactory;
-import hudson.widgets.buildsearch.BuildSearchParams;
+import jenkins.widgets.buildsearch.BuildSearchParamProcessor;
+import jenkins.widgets.buildsearch.BuildSearchParamProcessorFactory;
+import jenkins.widgets.buildsearch.BuildSearchParams;
 
 import java.util.List;
 
 /**
- * Search build history by {@link hudson.model.Run} description.
+ * Search build history by {@link hudson.model.Queue.Item} or {@link hudson.model.Run} name.
  *
  * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
  */
-public class DescriptionProcessorFactory extends BuildSearchParamProcessorFactory {
+public class NameProcessorFactory extends BuildSearchParamProcessorFactory {
 
     /**
-     * "desc" (description) search term.
+     * "name" search term.
      */
-    public static final String DESC_ST = "desc";
+    public static final String NAME_ST = "name";
 
-    private static final String[] SEARCH_TERMS = new String[] {DESC_ST}; // Build description
+    private static final String[] SEARCH_TERMS = new String[] {NAME_ST}; // Build name
 
     /**
      * {@inheritDoc}
@@ -58,20 +58,20 @@ public class DescriptionProcessorFactory extends BuildSearchParamProcessorFactor
      */
     @Override
     public BuildSearchParamProcessor createProcessor(BuildSearchParams searchParams) {
-        final List<BuildSearchParams.BuildSearchParam> descParams = searchParams.getParams(DESC_ST);
+        final List<BuildSearchParams.BuildSearchParam> nameParams = searchParams.getParams(NAME_ST);
 
-        if (descParams.isEmpty()) {
-            // "desc" search term not specified in search
+        if (nameParams.isEmpty()) {
+            // "name" search term not specified in search
             return null;
         }
 
         return new BuildSearchParamProcessor<String>() {
 
             @Override
-            public boolean fitsSearchParams(String description) {
-                // It fits if it contains any of the specified "desc" search terms.
-                for (BuildSearchParams.BuildSearchParam nameParam : descParams) {
-                    if (description.contains(nameParam.get())) {
+            public boolean fitsSearchParams(String name) {
+                // It fits if it contains any of the specified "name" search terms.
+                for (BuildSearchParams.BuildSearchParam nameParam : nameParams) {
+                    if (name.contains(nameParam.get())) {
                         return true;
                     }
                 }
@@ -80,12 +80,11 @@ public class DescriptionProcessorFactory extends BuildSearchParamProcessorFactor
 
             @Override
             public boolean fitsSearchParams(Queue.Item item) {
-                // Queue items don't have a description.
-                return false;
+                return fitsSearchParams(item.getDisplayName());
             }
             @Override
             public boolean fitsSearchParams(Run run) {
-                return fitsSearchParams(run.getDescription());
+                return fitsSearchParams(run.getDisplayName());
             }
         };
     }
