+++ core/src/test/java/jenkins/widgets/buildsearch/BuildSearchParamsTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package hudson.widgets.buildsearch;
+package jenkins.widgets.buildsearch;
 
 import org.junit.Assert;
 import org.junit.Test;
